feat(tecnico-service): adiciona metodo para buscar tecnico pelo id

+++ src/main/java/dev/joaorooliveira/help_desk/domain/tecnico/TecnicoRepository.java
@@ -3,5 +3,5 @@ package dev.joaorooliveira.help_desk.domain.tecnico;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface TecnicoRepository extends JpaRepository<Tecnico,Integer>, JpaSpecificationExecutor<Tecnico> {
+public interface TecnicoRepository extends JpaRepository<Tecnico,Long>, JpaSpecificationExecutor<Tecnico> {
 }
